backend/pyproject.toml
[project]
name = "prism-hack"
version = "0.1.0"
description = "Project for the Samsung Gen AI Hack"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.117.1",
    "liboqs-python",
    "piexif>=1.1.3",
    "pillow>=11.3.0",
    "python-multipart>=0.0.20",
    "torch>=2.8.0",
    "transformers>=4.56.2",
    "uvicorn>=0.37.0",
]

[tool.uv.sources]
liboqs-python = { git = "https://github.com/open-quantum-safe/liboqs-python.git" }
